// src/pages/2022/day02/assets/enums.tsx
export enum OpponentPlays {
    Rock = "A",
    Paper = "B",
    Scissors = "C",
}
export enum MyPlays {
    Rock = "X",
    Paper = "Y",
    Scissors = "Z",
}
export enum ResultLetter {
    Lose = "X",
    Draw = "Y",
    Win = "Z",
}
export enum MyPlaysPoints {
    Rock = 1,
    Paper = 2,
    Scissors = 3,
}
export enum RoundResultPoints {
    Win = 6,
    Lose = 0,
    Draw = 3,
}

export enum RoundResult {
    Win = 1,
    Lose = -1,
    Draw = 0,
}